LayoutTests/imported/w3c/web-platform-tests/custom-elements/parser/serializing-html-fragments-expected.txt


FAIL "is" value should be serialized if the custom element has no "is" content attribute The superclass is not a constructor.
FAIL "is" value should be serialized even for an undefined element assert_equals: expected "<p is=\"your-p\"></p>" but got "<p></p>"
FAIL "is" content attribute should be serialized even if the element is a customized built-in element The superclass is not a constructor.
